backend/requirements.txt
certifi==2022.9.24
interchange==2021.0.4
monotonic==1.6
packaging==21.3
pansi==2020.7.3
py2neo==2021.2.3
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.6
six==1.16.0
urllib3==1.26.12
Django==2.2.24
gunicorn==20.0.4
django-neomodel==0.0.7
djangorestframework==3.11.2
neo4j-driver==4.3.1
neomodel==4.0.4
pytz==2021.1
whitenoise==3.3.1
django-environ==0.8.1